Move last parameters function into utils of emutools
